+++ config/honeypot.php
@@ -43,4 +43,34 @@ return [
      * This switch determines if the honeypot protection should be activated.
      */
     'enabled' => env('HONEYPOT_ENABLED', true),
+
+    /*
+     * Response content returned when a request is determined to be spam.
+     */
+    'response_content' => "If you're a robot, you've been caught by a human.  If you're a human, another human has mistaken you for a robot.",
+
+    /*
+     * Enable Honeypot traps & specify input names.
+     */
+    'traps' => [
+        /*
+         * Catch spam requests that submit an identical 'first' & 'last' name inputs.
+         */
+        'duplicate_names' => [
+            /*
+             * This switch determines if the 'duplicate_names' trap is enabled
+             */
+            'enabled' => false,
+
+            /*
+             * Name of the 'first name' attribute used in Honeypot request
+             */
+            'name_first_input' => 'data.name_first',
+
+            /*
+             * Name of the 'last name' attribute used in Honeypot request
+             */
+            'name_last_input' => 'data.name_last',
+        ],
+    ],
 ];