+++ tests/TestCase.php
@@ -14,6 +14,13 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
     use RefreshDatabase;
 
     /**
+     * Indicates whether the default seeder should run before each test.
+     *
+     * @var bool
+     */
+    protected $seed = false;
+
+    /**
      * Register package service providers.
      *
      * @param Application $app
@@ -39,6 +46,11 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
         if (! defined('LARAVEL_START')) {
             define('LARAVEL_START', microtime(true));
         }
+
+        // Set config values
+        $app['config']->set('tracking.traffic.track', true);
+        $app['config']->set('tracking.driver', 'sync');
+        $app['config']->set('honeypot.traps.duplicate_names.enabled', true);
 
         // Migrate 'track_traffic' table
         include_once __DIR__.'/../vendor/sfneal/tracking/database/migrations/create_track_traffic_table.php.stub';